MultiModal_RAG/requirements.txt
fitz
transformers
langchain
pillow
torch
numpy
scikit-learn
langchain[openai]
faiss-cpu